Refac : foundations styled 분리
