+++ src/foundations/Icon/Icon.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
-import styled from "styled-components";
 
+import { Svg } from "./Icon.styled";
 import { icons } from "@/_shared";
 
 const Icon = ({ icon, block, ...props }) => {
@@ -31,14 +31,3 @@ Icon.defaultProps = {
 };
 
 export default Icon;
-
-const Svg = styled.svg`
-  display: ${(props) => (props.block ? "block" : "inline-block")};
-  vertical-align: middle;
-  shape-rendering: inherit;
-  transform: translate3d(0, 0, 0);
-
-  stroke-width: 2;
-  stroke-linecap: round;
-  stroke-linejoin: round;
-`;
